requirements.txt
psycopg2==2.9.3
PyYAML==6.0
Scrapy==2.6.2